+++ src/modules/tools/routing/store/directions/stateDirections.js
@@ -44,6 +44,7 @@ export default {
     // Directions Parameter
     waypoints: [],
     routingAvoidFeaturesOptions: [],
+    deleteRoutes: true,
     // Routing Directions Result
     routingDirections: null,
     mapListenerAdded: false,